+++ src/main/java/com/example/recruit_page_wwy/HelloController.java
@@ -127,14 +127,4 @@ public class HelloController {
         return "redirect:/";
     }
 
-    @PostMapping("/login")
-    public String login() {
-        return "redirect:/";
-    }
-
-    @PostMapping("/join")
-    public String join() {
-        return "redirect:/login-form";
-    }
-
 }
