+++ src/main/java/com/example/recruit_page_wwy/user/UserController.java
@@ -3,6 +3,7 @@ package com.example.recruit_page_wwy.user;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 
 @RequiredArgsConstructor
 @Controller
@@ -19,8 +20,15 @@ public class UserController {
     // UserJoinPage
     @GetMapping("/join-form/user")
     public String userJoinForm(UserRequest.UserDTO reqDTO) {
-        userService.joinUser(reqDTO);
         return "user/user-join-form";
+    }
+
+    // UserJoin
+    @PostMapping("/join")
+    public String userJoin(UserRequest.UserDTO reqDTO) {
+        userService.joinUser(reqDTO);
+        System.out.println(reqDTO);
+        return "redirect:/login-form";
     }
 
     // ComJoinPage
@@ -28,6 +36,11 @@ public class UserController {
     public String comJoinForm(UserRequest.ComDTO reqDTO) {
         userService.joinCom(reqDTO);
         return "user/com-join-form";
+    }
+
+    @PostMapping("/login")
+    public String login() {
+        return "redirect:/";
     }
 
 }
