+++ src/main/resources/application.properties
@@ -15,7 +15,7 @@ spring.jpa.hibernate.ddl-auto=create
 # query log
 spring.jpa.show-sql=true
 # dummy data
-#spring.sql.init.data-locations=classpath:db/data.sql
+spring.sql.init.data-locations=classpath:db/data.sql
 # create dummy data  after ddl-auto create
 spring.jpa.defer-datasource-initialization=true
 # mustache request expose
