+++ src/main/resources/static/css/user-save-form.css
@@ -17,6 +17,7 @@ body {
     align-items: center;
 }
 
+
 .felx-box {
     display: flex;
     justify-content: center;
@@ -33,17 +34,17 @@ body {
     text-align: center;
 }
 
-input[type="text"] {
+
+.input-box {
+    position: relative;
+}
+
+.input-box > input {
     width: 100%;
     height: 36px;
     border-radius: 8px;
     border: 1px solid #ccc;
     padding-left: 10px;
-}
-
-.input-box {
-
-    position: relative;
 }
 
 .input-box > button {
