+++ src/main/resources/templates/user/user-join-form.mustache
@@ -1,29 +1,29 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-	<meta charset="UTF-8"/>
-	<meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-	<title>구직자 회원가입</title>
-	<link
-			rel="stylesheet"
-			href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
-	/>
-	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
-	<link
-			href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
-			rel="stylesheet"
-	/>
-	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-	<link
-			rel="stylesheet"
-			href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css"
-	/>
+    <meta charset="UTF-8"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+    <title>구직자 회원가입</title>
+    <link
+            rel="stylesheet"
+            href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
+    />
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
+    <link
+            href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
+            rel="stylesheet"
+    />
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
+    <link
+            rel="stylesheet"
+            href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css"
+    />
 
-	<link rel="stylesheet" href="../css/user-save-form.css"/>
-	<link
-			href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR&display=swap"
-			rel="stylesheet"
-	/>
+    <link rel="stylesheet" href="../css/user-save-form.css"/>
+    <link
+            href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR&display=swap"
+            rel="stylesheet"
+    />
 </head>
 
 <!-- nav -->
@@ -31,46 +31,48 @@
 
 <!-- main -->
 <form action="/join" method="post">
-	<div class="felx-box">
-		<div class="form-box">
-			<div class="save">구직자 회원가입</div>
+    <div class="felx-box">
+        <div class="form-box">
+            <div class="save">구직자 회원가입</div>
 
-			<div>
-				<div class="input-header">아이디</div>
-				<div class="input-box">
-					<input type="text" placeholder="아이디를 입력해주세요"/>
-					<div class="button-box">
-						<button type="button" class="check-box">중복확인</button>
-					</div>
-				</div>
-			</div>
+            <div>
+                <div class="input-header">이메일</div>
+                <div class="input-box">
+                    <input type="email" placeholder="이메일을 입력해주세요" name="email"/>
+                    <div class="button-box">
+                        <button type="button" class="check-box">중복확인</button>
+                    </div>
+                </div>
+            </div>
 
-			<div>
-				<div class="input-header">이메일</div>
-				<input type="text" placeholder="이메일을 입력해주세요"/>
-			</div>
+            <div>
+                <div class="input-header">이름</div>
+                <div class="input-box">
+                    <input type="text" placeholder="이름을 입력해주세요" name="username"/>
+                </div>
+            </div>
 
-			<div>
-				<div class="input-header">휴대전화</div>
-				<input type="text" placeholder="휴대전화를 입력해주세요"/>
-			</div>
+            <div>
+                <div class="input-header">휴대전화</div>
+                <input type="text" placeholder="휴대전화를 입력해주세요" name="phone"/>
+            </div>
 
-			<div>
-				<div class="input-header">비밀번호</div>
-				<input type="text" placeholder="비밀번호를 입력해주세요"/>
-				<div class="input-header">비밀번호확인</div>
-				<input type="text" placeholder="비밀번호를 재입력해주세요"/>
-				<div class="checkpwd">
-					영문 대소문자, 숫자, 특수문자를 3가지 이상으로 조합해 8자
-					이상 16자 이하로 입력해주세요.
-				</div>
-			</div>
+            <div>
+                <div class="input-header">비밀번호</div>
+                <input type="password" placeholder="비밀번호를 입력해주세요" name="password"/>
+                <div class="input-header">비밀번호확인</div>
+                <input type="password" placeholder="비밀번호를 재입력해주세요" name="password"/>
+                <div class="checkpwd">
+                    영문 대소문자, 숫자, 특수문자를 3가지 이상으로 조합해 8자
+                    이상 16자 이하로 입력해주세요.
+                </div>
+            </div>
 
-			<div class="save-box">
-				<button class="save-btn">가입하기</button>
-			</div>
-		</div>
-	</div>
+            <div class="save-box">
+                <button class="save-btn">가입하기</button>
+            </div>
+        </div>
+    </div>
 </form>
 
 <!-- footer -->
